pyproject.toml
[project]
name = "U1bot"
version = "0.1.0"
description = "U1bot"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.poetry]
name = "u1bot"
version = "0.1.0"
description = ""
authors = ["wling-art <wlingzhenyu@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = {extras = ["fastapi"], version = "^2.1.3"}
sqlalchemy = "^2.0.24"
tortoise = "^0.1.1"
nonebot-plugin-tortoise-orm = "^0.1.1"
ujson = "^5.9.0"
nonebot-plugin-apscheduler = "^0.3.0"
httpx = "^0.26.0"
pil-utils = "^0.1.8"
faker = "^22.0.0"
requests = "^2.31.0"
nonebot-plugin-htmlrender = "^0.2.3"
nonebot-plugin-addfriend = "^2.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "ali"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
priority = "primary"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "default"